+++ src/main/java/pl/asie/computronics/block/BlockCamera.java
@@ -1,13 +1,16 @@
 package pl.asie.computronics.block;
 
+import cpw.mods.fml.common.Optional;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
+import li.cil.oc.api.network.Environment;
 import net.minecraft.client.renderer.texture.IIconRegister;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.IIcon;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import pl.asie.computronics.reference.Config;
+import pl.asie.computronics.reference.Mods;
 import pl.asie.computronics.tile.TileCamera;
 
 public class BlockCamera extends BlockMachineSidedIcon {
@@ -40,5 +43,11 @@ public class BlockCamera extends BlockMachineSidedIcon {
 	@Override
 	public boolean emitsRedstone(IBlockAccess world, int x, int y, int z, int side) {
 		return Config.REDSTONE_REFRESH;
+	}
+
+	@Override
+	@Optional.Method(modid= Mods.OpenComputers)
+	public Class<? extends Environment> getTileEntityClass(int meta) {
+		return TileCamera.class;
 	}
 }