+++ src/main/java/pl/asie/computronics/block/BlockDigitalDetector.java
@@ -1,7 +1,9 @@
 package pl.asie.computronics.block;
 
+import cpw.mods.fml.common.Optional;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
+import li.cil.oc.api.network.Environment;
 import mods.railcraft.client.util.textures.TextureAtlasSheet;
 import mods.railcraft.common.blocks.tracks.TrackTools;
 import mods.railcraft.common.items.IActivationBlockingItem;
@@ -20,6 +22,7 @@ import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.common.util.ForgeDirection;
 import pl.asie.computronics.Computronics;
+import pl.asie.computronics.reference.Mods;
 import pl.asie.computronics.tile.TileDigitalDetector;
 
 /**
@@ -210,5 +213,11 @@ public class BlockDigitalDetector extends BlockPeripheral {
 	@Override
 	public boolean canConnectRedstone(IBlockAccess world, int i, int j, int k, int dir) {
 		return false;
+	}
+
+	@Override
+	@Optional.Method(modid= Mods.OpenComputers)
+	public Class<? extends Environment> getTileEntityClass(int meta) {
+		return TileDigitalDetector.class;
 	}
 }