+++ src/main/java/pl/asie/computronics/block/BlockIronNote.java
@@ -2,6 +2,7 @@ package pl.asie.computronics.block;
 
 import cpw.mods.fml.common.Loader;
 import cpw.mods.fml.common.Optional;
+import li.cil.oc.api.network.Environment;
 import net.minecraft.block.Block;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.world.World;
@@ -56,5 +57,11 @@ public class BlockIronNote extends BlockPeripheral implements IRedNetInputNode {
 				NoteUtils.playNote(world, x, y, z, -1, i);
 			inputValue >>= 1;
 		}
+	}
+
+	@Override
+	@Optional.Method(modid= Mods.OpenComputers)
+	public Class<? extends Environment> getTileEntityClass(int meta) {
+		return TileIronNote.class;
 	}
 }