+++ src/main/java/pl/asie/computronics/block/BlockLocomotiveRelay.java
@@ -1,11 +1,14 @@
 package pl.asie.computronics.block;
 
+import cpw.mods.fml.common.Optional;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
+import li.cil.oc.api.network.Environment;
 import net.minecraft.client.renderer.texture.IIconRegister;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.IIcon;
 import net.minecraft.world.World;
+import pl.asie.computronics.reference.Mods;
 import pl.asie.computronics.tile.TileLocomotiveRelay;
 
 /**
@@ -45,5 +48,11 @@ public class BlockLocomotiveRelay extends BlockPeripheral {
 			default:
 				return mSide;
 		}
+	}
+
+	@Override
+	@Optional.Method(modid= Mods.OpenComputers)
+	public Class<? extends Environment> getTileEntityClass(int meta) {
+		return TileLocomotiveRelay.class;
 	}
 }